Fixed MO unit test to import paddle conditionally.

+++ tools/mo/unit_tests/mo/utils/mo_fallback_test_actual.py
@@ -4,7 +4,6 @@
 import unittest
 from unittest.mock import patch, Mock
 
-import openvino
 from openvino.tools.mo.convert_impl import prepare_ir
 from openvino.tools.mo.utils.error import Error
 from openvino.frontend import FrontEndManager, FrontEnd # pylint: disable=no-name-in-module,import-error
@@ -12,7 +11,6 @@ from onnx.helper import make_graph, make_model, make_tensor_value_info
 import argparse
 import os
 import onnx
-import paddle
 import numpy as np
 import shutil
 import pytest
@@ -24,6 +22,11 @@ try:
 except ImportError:
     import openvino.tools.mo.utils.telemetry_stub as tm
 
+try:
+    import paddle
+    paddle_imported = True
+except ImportError:
+    paddle_imported = False
 
 def base_args_config(use_legacy_fe:bool=None, use_new_fe:bool=None):
     args = argparse.Namespace()
@@ -121,19 +124,20 @@ class TestMoFallback(unittest.TestCase):
             with open(file, 'w') as f:
                 f.write(content)
 
-        self.paddle_dir = "paddle_dir"
-        paddle.enable_static()
-        if not os.path.exists(self.paddle_dir):
-            os.mkdir(self.paddle_dir)
-        x = np.array([-2, 0, 1]).astype('float32')
-        node_x = paddle.static.data(name='x', shape=x.shape, dtype='float32')
-        out = paddle.nn.functional.relu(node_x)
+        if paddle_imported:
+            self.paddle_dir = "paddle_dir"
+            paddle.enable_static()
+            if not os.path.exists(self.paddle_dir):
+                os.mkdir(self.paddle_dir)
+            x = np.array([-2, 0, 1]).astype('float32')
+            node_x = paddle.static.data(name='x', shape=x.shape, dtype='float32')
+            out = paddle.nn.functional.relu(node_x)
 
-        cpu = paddle.static.cpu_places(1)
-        exe = paddle.static.Executor(cpu[0])
-        exe.run(paddle.static.default_startup_program())
+            cpu = paddle.static.cpu_places(1)
+            exe = paddle.static.Executor(cpu[0])
+            exe.run(paddle.static.default_startup_program())
 
-        save_paddle_model("relu", exe, feedkeys=['x'], fetchlist=[out], target_dir=self.paddle_dir)
+            save_paddle_model("relu", exe, feedkeys=['x'], fetchlist=[out], target_dir=self.paddle_dir)
 
 
     def tearDown(self):
@@ -141,7 +145,8 @@ class TestMoFallback(unittest.TestCase):
             os.remove(name)
         for name in self.test_config_files:
             os.remove(name)
-        shutil.rmtree(self.paddle_dir)
+        if paddle_imported:
+            shutil.rmtree(self.paddle_dir)
 
 
     @generate(*[(['dir_to_extension'], None, None, 'mo_legacy', 'extensions'), # fallback
@@ -278,6 +283,8 @@ class TestMoFallback(unittest.TestCase):
                 (None, None, None, 'paddle_frontend'),
     ])
     def test_no_fallback_if_pdpd(self, use_new_fe, use_legacy, extension, expected_path):
+        if not paddle_imported:
+            return
         args = base_args_config(use_legacy, use_new_fe)
         args.framework = 'paddle'
         args.extensions = extension
@@ -291,6 +298,8 @@ class TestMoFallback(unittest.TestCase):
 
 
     def test_exception_if_old_extensions_used_for_pdpd(self):
+        if not paddle_imported:
+            return
         args = base_args_config()
         args.framework = 'paddle'
         args.extensions = ['dir_to_extension']
